fix: tab focus is tracked in menubar instead of menuitem
